requirements.txt
crewai>=0.80.0
crewai-tools>=0.15.0
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0
python-dotenv>=1.0.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
requests>=2.32.0
pydantic>=2.10.0
pyyaml>=6.0.0
pytest>=8.0.0
pandas>=2.2.0
plotly>=5.24.0
